Refactor project dir structure
